+++ package.json
@@ -16,6 +16,7 @@
     "@radix-ui/react-select": "2.1.4",
     "@radix-ui/react-slot": "1.1.1",
     "class-variance-authority": "^0.7.1",
+    "date-fns": "^3.6.0",
     "clsx": "^2.1.1",
     "firebase": "^11.7.3",
     "lucide-react": "^0.454.0",
